+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/snap/SnapToRouteTest.java
@@ -15,6 +15,7 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import static junit.framework.Assert.assertEquals;
+import static org.mockito.internal.matchers.text.ValuePrinter.print;
 
 @RunWith(RobolectricTestRunner.class)
 public class SnapToRouteTest extends BaseTest {
@@ -171,6 +172,15 @@ public class SnapToRouteTest extends BaseTest {
     location.setLongitude(-77.0782341);
     location.setBearing(20);
 
+    Location previousSnappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+            routeProgress,
+            0.8,
+            0.8,
+            200,
+            20,
+            0
+    ));
+
     Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
             routeProgress,
             0.8,
@@ -180,7 +190,30 @@ public class SnapToRouteTest extends BaseTest {
             0
     ));
 
-    assertEquals(358.19876f, snappedLocation.getBearing());
+    // Latest snapped bearing should be used, because next lef is not containing enough steps
+    assertEquals(previousSnappedLocation.getBearing(), snappedLocation.getBearing());
+  }
+
+  @Test
+  public void getSnappedLocation_bearingNoBearingBeforeWithSingleStepLegBeforeNextLeg() throws Exception {
+    DirectionsRoute routeProgress = buildMultipleLegRoute(SINGLE_STEP_LEG);
+    Snap snap = new SnapToRoute();
+    Location location = new Location("test");
+    location.setLatitude(38.8943771);
+    location.setLongitude(-77.0782341);
+    location.setBearing(20);
+
+    Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+            routeProgress,
+            0.8,
+            0.8,
+            200,
+            21,
+            0
+    ));
+
+    // Fallback to location bearing if no previous bearing was calculated.
+    assertEquals(location.getBearing(), snappedLocation.getBearing());
   }
 
   @Test
@@ -192,6 +225,15 @@ public class SnapToRouteTest extends BaseTest {
     location.setLongitude(-77.0282631);
     location.setBearing(20);
 
+    Location lastSnappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
+        routeProgress,
+        0.6,
+        0.6,
+        0.6,
+        14,
+        1
+    ));
+
     Location snappedLocation = snap.getSnappedLocation(location, buildTestRouteProgress(
         routeProgress,
         0.8,
@@ -201,7 +243,8 @@ public class SnapToRouteTest extends BaseTest {
         1
     ));
 
-    assertEquals(0f, snappedLocation.getBearing());
+    // Latest snapped bearing should be used, because no future steps are available
+    assertEquals(lastSnappedLocation.getBearing(), snappedLocation.getBearing());
   }
 
   private DirectionsRoute buildMultipleLegRoute() throws Exception {